# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the official data model for BARAFIRI API documentation.",
        "title": "Barafiri Backend API documentation",
        "termsOfService": "http://my-gpi.io/terms/",
        "contact": {
            "name": "Endy Apinageri",
            "url": "https://www.endyapina.com",
            "email": "endy.apinageri@my-gpi.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8112",
    "basePath": "/api/v1",
    "paths": {
...
        "/util/category/": {
            "get": {
                "description": "Get the list of all Categories",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public API"
                ],
                "summary": "Endpoint to get the list of all categories on the database",
                "responses": {
                    "200": {
                        "description": "Category list and sub category",
                        "schema": {
                            "$ref": "#/definitions/controller.SubCategoryList"
                        }
                    },
                    "400": {
                        "description": "Category list",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseObject"
                        }
                    }
                }
            }
        },
        "/util/industry/": {
            "get": {
                "description": "Get the list of all industries",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public API"
                ],
                "summary": "Endpoint to get the list of all industries on the database",
                "responses": {
                    "200": {
                        "description": "Category list and sub category",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/controller.Industries"
                            }
                        }
                    },
                    "400": {
                        "description": "Category list",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseObject"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "controller.Industries": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "industry": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "controller.ResponseObject": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "object"
                },
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "controller.SubCategoryList": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "subcategories": {
                    "type": "array",
                    "items": {
                        "type": "object"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
...
}